Improve custom tests
